+++ src/Backends/Tasks/Task.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Rubix\ML\Backends\Tasks;
+
+use Rubix\ML\Deferred;
+
+class Task extends Deferred
+{
+    //
+}