+++ tests/Backends/AmpTest.php
@@ -2,9 +2,9 @@
 
 namespace Rubix\ML\Tests\Backends;
 
-use Rubix\ML\Deferred;
 use Rubix\ML\Backends\Amp;
 use Rubix\ML\Backends\Backend;
+use Rubix\ML\Backends\Tasks\Task;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -48,10 +48,8 @@ class AmpTest extends TestCase
      */
     public function enqueueProcess() : void
     {
-        $functions = array_fill(0, 10, [self::class, 'foo']);
-
-        foreach ($functions as $i => $function) {
-            $this->backend->enqueue(new Deferred($function, [$i]));
+        for ($i = 0; $i < 10; ++$i) {
+            $this->backend->enqueue(new Task([self::class, 'foo'], [$i]));
         }
 
         $results = $this->backend->process();
@@ -65,6 +63,6 @@ class AmpTest extends TestCase
      */
     public static function foo(int $i) : array
     {
-        return [$i, microtime(true)];
+        return [$i * 2, microtime(true)];
     }
 }